Clean up confirmation flow
